tkey, fido2: fix implicit integer conversion warnings
Add explicit casts to silence implicit conversion warnings in
attestation.c, device.c, fifo.c, and fs.c. Change loop and count
variables from int/uint16_t to size_t where appropriate, and cast
lfs_file_seek() offset arguments to lfs_soff_t.

Fix memcpy-based cert_size assignment in attestation_read_cert() with
a proper assignment, moved to the success path only. Add missing error
check for fs_file_size() return value in ctap_open_rk_file(). Change
ctap_load_rk() index parameter from int to uint8_t in device.h and
device.c to match its callers.

In ctap_credential_management.c, add a bounds check against
ctap_max_number_of_rks() before using the RK count, and narrow
nbr_rk and index to uint8_t accordingly.

+++ fido2/inc/device.h
@@ -180,7 +180,7 @@ int ctap_close_rk_file(void);
  *
  * *Optional*, if not implemented, operates on non-persistant RK's.
  */
-void ctap_load_rk(int index, CTAP_residentKey *dst_rk);
+void ctap_load_rk(uint8_t index, CTAP_residentKey *dst_rk);
 
 /** Read the next resident key from already open file into memory
  * @param dst_rk pointer to resident key structure to write into with RK.
